os/setthreadpri(.data) -> os/thread
