+++ src/os/thread.c
@@ -1,4 +1,11 @@
+#include "ultra64.h"
 #include "PR/osint.h"
+
+struct __osThreadTail __osThreadTail = {0, -1};
+OSThread* __osRunQueue = (OSThread*) &__osThreadTail;
+OSThread* __osActiveQueue = (OSThread*) &__osThreadTail;
+OSThread* __osRunningThread = NULL;
+OSThread* __osFaultedThread = NULL;
 
 void __osDequeueThread(OSThread **queue, OSThread *t) {
     register OSThread *pred;
